Correct tests for valid_variants file
