+++ Sources/VariantsCore/Custom Types/Project/AndroidProject.swift
@@ -40,7 +40,7 @@ class AndroidProject: Project {
         }
 
         guard let desiredVariant = configuration.variants.first(where: { $0.name.lowercased() == variant.lowercased() }) else {
-            throw ValidationError("Variant \(variant) not found.")
+            throw ValidationError("Variant '\(variant)' not found.")
         }
 
         do {
@@ -176,8 +176,8 @@ class AndroidProject: Project {
     // swiftlint:enable function_body_length
     
     private func storeFastlaneParams(_ parameters: [CustomProperty], configuration: AndroidConfiguration) throws {
-        let fastlaneParamPath = try Path(configuration.path).safeJoin(path: StaticPath.Fastlane.parametersFolder)
-        try fastlaneFactory.createParametersFile(in: fastlaneParamPath, with: parameters)
+         
+        try fastlaneFactory.createParametersFile(in: StaticPath.Fastlane.parametersFolder, with: parameters)
     }
     
     private let gradleFactory: GradleFactory
